Fix middleware activation order
The mock server middleware ignores everything done after it. If we make
a proxy with a mock server and no port, the DefaultParams which carries
the port gets set after the mock middleware, and then we're in a bit of
a bind. So, instead of immediately setting the middleware during client
creation when we may not have a port if the user doesn't specify one, we
delay a bit longer until we definitely have a port.

That way, when we _set_middlewares, we can properly set our
DefaultParams middleware before the mock server gets activated.